Personaje no sale por pantalla pero casi
Muchos includes y no da errores, pero no sale el personaje por pantalla

+++ 00ProyectoCrypt/Timer.cpp
@@ -27,7 +27,7 @@ void Timer::stop()
 	}
 }
 
-uint32_t Timer::getTime()
+float Timer::getTime()
 {
 	return SDL_GetTicks();
 }
